+++ src/components/channels/Index.vue
@@ -3,9 +3,8 @@
     <template  v-if="!active">
       <div class="text-center" style="display: flex; justify-content: center; font-size: 1.5rem">
         <div class="text-grey-3" style="margin-top: 50px">
-          Select channel
           <q-btn flat style="display: flex; flex-wrap: nowrap; margin-top: 20px" icon-right="mdi-menu-down" v-if="items.length">
-            &lt;Available channels&gt;
+            Select channel
             <q-popover fit ref="popoverNotActive">
               <q-list link separator class="scroll">
                 <q-item
@@ -23,6 +22,7 @@
               </q-list>
             </q-popover>
           </q-btn>
+          <div v-if="!items.length">Channels is empty</div>
         </div>
       </div>
     </template>
@@ -136,12 +136,8 @@
     methods: {
       settingsClickHandler () {
         Dialog.create({
-          title: 'Settings',
+          title: 'Ratio',
           form: {
-            ratioHeader: {
-              type: 'heading',
-              label: 'Ratio'
-            },
             ratio: {
               type: 'radio',
               model: this.ratio,