+++ src/components/channels/messages/Index.vue
@@ -12,6 +12,7 @@
       :i18n="i18n"
       :filter="filter"
       :theme="theme"
+      :title="'Messages'"
       @change:filter="filterChangeHandler"
       @change:pagination-next="paginationNextChangeHandler"
       @change:mode="modeChange"
@@ -63,7 +64,7 @@
       messages: {
         get () {
           let messages = this.$store.state[this.moduleName].messages
-          this.i18n.to = messages.length ? date.formatDate(this.from * 1000, 'HH:mm:ss') : 'Next'
+          this.i18n.to = messages.length ? `Next batch from ${date.formatDate(this.from * 1000, 'HH:mm:ss')}` : 'Next'
           return this.filter ? this.filterMessages(this.filter, messages) : messages
         },
         set (val) {